(sista) uppdateringen
flyttade logik från controllern till handlers
städning och refactoring

+++ src/main/java/com/example/Dist_sys_lab1_webshop/Model/Item/CategoryHandler.java
@@ -8,10 +8,15 @@ public class CategoryHandler {
 
     
     public static void addCategory(String category){
-        CategoryDB.addCategory(category);
+        if (category != null) {
+            CategoryDB.addCategory(category);
+        }
     }
 
-    public static void removeCategoryById(int id){
+    public static void removeCategoryById(String categoryId){
+        if (categoryId == null) return;
+
+        int id = Integer.parseInt(categoryId);
         CategoryDB.removeCategoryById(id);
     }
     public static ArrayList<Category> getCategories() {
